requirements.txt
yacs==0.1.8
numpy==1.22.3
torchvision==0.9.0+cu111
tqdm==4.66.1
opencv_contrib_python==4.5.3.56
imageio==2.9.0
torch==1.8.0+cu111
mmcv_full==1.3.10
matplotlib==3.4.3
mmcv==1.3.17
Pillow==8.1.1
scikit_learn==0.24.1

